Add hash table files
I'm a bit too tired to think through how I want to structure the
implementation though. I figured I'd use an array of linked lists, but I
forgot to add something that will connect the hash value to a bucket.
Though maybe it's the caller's responsibility to ensure that the hash
function produces a value in the range of 0 - bucket_count.